pyproject.toml
[project]
name = "long-running-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.12.0",
    "openai>=2.8.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
]
